restore shifts to pre-rounding significand tracking

+++ softfloat/s_roundPackToBF16.c
@@ -54,7 +54,7 @@ bfloat16_t
     *------------------------------------------------------------------------*/
     softfloat_intermediateResult.sign     = sign;
     softfloat_intermediateResult.exp      = exp;
-    softfloat_intermediateResult.sig64    = ((uint64_t)sig)<<48;
+    softfloat_intermediateResult.sig64    = ((uint64_t)sig) << (uint64_t)48ULL;
     softfloat_intermediateResult.sig0     = 0;
     softfloat_intermediateResult.sigExtra = 0;
